refactor(share): split share.rs into compression, devices, fetch, send, and types modules
