+++ src/fulgur/sync/share/mod.rs
@@ -1,0 +1,11 @@
+pub(super) mod compression;
+pub(super) mod devices;
+pub(super) mod fetch;
+pub(super) mod send;
+pub(super) mod types;
+
+pub use compression::decompress_content;
+pub use devices::{Device, get_devices, get_icon};
+pub use fetch::fetch_pending_shares;
+pub use send::{MAX_SYNC_SHARE_PAYLOAD_BYTES, share_file};
+pub use types::{ShareFileRequest, ShareResult};
